+++ src/DearModdingUI/controls/Faq.cpp
@@ -37,8 +37,7 @@ namespace DearModdingUI
 					.expanded = &expanded->second,
 					.textColor = textColor,
 					.hoveredTextColor = textColor,
-					.highlightStyle = RowHighlightStyle::kRoundedFill,
-					.clickBehavior = RowClickBehavior::kToggle
+					.highlightStyle = RowHighlightStyle::kRoundedFill
 				});
 			}
 			if (expanded->second)